Resolved conflicts on prod

+++ EPlast/EPlast.BLL/Services/Blank/AchievementDocumentService.cs
@@ -15,16 +15,19 @@ namespace EPlast.BLL.Services.Blank
     public class AchievementDocumentService : IBlankAchievementDocumentService
     {
         private readonly IRepositoryWrapper _repositoryWrapper;
+        private readonly IUserCourseService _usercourseService;
         private readonly IMapper _mapper;
         private readonly IBlankAchievementBlobStorageRepository _blobStorageRepo;
 
         public AchievementDocumentService(IRepositoryWrapper repositoryWrapper,
            IMapper mapper,
-           IBlankAchievementBlobStorageRepository blobStorageRepo)
+           IBlankAchievementBlobStorageRepository blobStorageRepo,
+           IUserCourseService usercourseService)
         {
             _repositoryWrapper = repositoryWrapper;
             _mapper = mapper;
             _blobStorageRepo = blobStorageRepo;
+            _usercourseService = usercourseService;
         }
 
         public async Task<IEnumerable<AchievementDocumentsDto>> AddDocumentAsync(IEnumerable<AchievementDocumentsDto> achievementDocumentsDTO)
@@ -46,15 +49,16 @@ namespace EPlast.BLL.Services.Blank
             return achievementDocumentsDTO;
         }
 
-        public async Task DeleteFileAsync(int documentId)
+        public async Task DeleteFileAsync(int documentId, string userId)
         {
             var document = await _repositoryWrapper.AchievementDocumentsRepository
                 .GetFirstOrDefaultAsync(d => d.ID == documentId);
-
-            await _blobStorageRepo.DeleteBlobAsync(document.BlobName);
-
-            _repositoryWrapper.AchievementDocumentsRepository.Delete(document);
-            await _repositoryWrapper.SaveAsync();
+            if (document != null)
+            {
+                await _blobStorageRepo.DeleteBlobAsync(document.BlobName);
+                _repositoryWrapper.AchievementDocumentsRepository.Delete(document);
+                await _repositoryWrapper.SaveAsync();
+            }
         }
 
         public async Task<string> DownloadFileAsync(string fileName)
@@ -62,18 +66,24 @@ namespace EPlast.BLL.Services.Blank
             return await _blobStorageRepo.GetBlobBase64Async(fileName);
         }
 
-        public async Task<IEnumerable<AchievementDocumentsDto>> GetDocumentsByUserIdAsync(string userid)
+        public async Task<IEnumerable<AchievementDocumentsDto>> GetDocumentsByUserIdAsync(string userId)
         {
             return _mapper.Map<IEnumerable<AchievementDocuments>, IEnumerable<AchievementDocumentsDto>>(
-                await _repositoryWrapper.AchievementDocumentsRepository.GetAllAsync(i => i.UserId == userid));
+                await _repositoryWrapper.AchievementDocumentsRepository.GetAllAsync(i => i.UserId == userId));
         }
 
-        public async Task<IEnumerable<AchievementDocumentsDto>> GetPartOfAchievementAsync(int pageNumber, int pageSize, string userid)
+        public async Task<IEnumerable<AchievementDocumentsDto>> GetPartOfAchievementByUserIdAsync(int pageNumber, int pageSize, string userId)
         {
-            var partOfAchievements = (await _repositoryWrapper.AchievementDocumentsRepository.GetAllAsync(i => i.UserId == userid)).Skip(pageSize * pageNumber).Take(pageSize);
-
+            var achievements = await _repositoryWrapper.AchievementDocumentsRepository.GetAllAsync(i => i.UserId == userId);
+            var partOfAchievements = achievements.Skip(pageSize * pageNumber).Take(pageSize);
             return _mapper.Map<IEnumerable<AchievementDocuments>, IEnumerable<AchievementDocumentsDto>>(partOfAchievements);
-            
+        }
+
+        public async Task<IEnumerable<AchievementDocumentsDto>> GetPartOfAchievementByUserIdAndCourseIdAsync(int pageNumber, int pageSize, string userId, int courseId)
+        {
+            var achievements = await _repositoryWrapper.AchievementDocumentsRepository.GetAllAsync(i => i.UserId == userId && i.CourseId == courseId);
+            var partOfAchievements = achievements.Skip(pageSize * pageNumber).Take(pageSize);
+            return _mapper.Map<IEnumerable<AchievementDocuments>, IEnumerable<AchievementDocumentsDto>>(partOfAchievements);
         }
     }
 }
